+++ back-end/foodiefair/src/main/java/com/multicampus/foodiefair/dto/CommentDTO.java
@@ -6,9 +6,9 @@ import java.util.Date;
 
 @Data
 public class CommentDTO {
-    private Integer commentId;      // 댓글ID
-    private Integer userId;         // 회원ID
-    private Integer reviewId;       // 리뷰ID
+    private long commentId;      // 댓글ID
+    private int userId;         // 회원ID
+    private long reviewId;       // 리뷰ID
     private String commentContent;  // 댓글내용
     private Date commentDate;       // 작성날짜
 }
