+++ back-end/foodiefair/src/main/java/com/multicampus/foodiefair/dto/QuestionDTO.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @NoArgsConstructor
 @ToString
 public class QuestionDTO {
-    private int questionId; //문의id
+    private long questionId; //문의id
     private int userId; //회원id
     private String questionType; //문의 종류
     private Date questionDate; //문의 시간
