+++ back-end/foodiefair/src/main/java/com/multicampus/foodiefair/service/IUserService.java
@@ -9,6 +9,6 @@ public interface IUserService {
     public UserDTO getUserByEmail(String userEmail);
     public UserDTO getUserById(Long id);
     public void updateUser(UserDTO userDto);
-    public void deleteUser(Long userId);
+    public void deleteUser(int userId);
     public void updateUserPassword(String userEmail, String userPwd);
 }
